+++ workspace/currency_exchange/test/__init__.py
@@ -1,0 +1,2 @@
+"""Test package for the currency exchange task workspace."""
+